Add an option to try and disambiguate when more than one eligible key is found via JwksDecryptionKeyResolver by attempting decryption with each of the candidate keys. Useful, for example, when you have more than one eligible key of the same 'kind' (for rotation maybe) but the sender isn't including the kid in the JWE.
Add a selectWithAttemptDecryptDisambiguate(...) to DecryptionJwkSelector that will try an actual decryption with all selected keys from selectList(...) when there is more than one that was otherwise selected based on the headers. And add a setDisambiguateWithAttemptDecrypt(true) option on JwksDecryptionKeyResolver to use that rather than the select(...) that gives one key based on the header selection and just the first one if more than one fit the selection.